refactor(web): extract Telegram helpers into src/web/telegram.ts
Move all Telegram bot API interaction out of web.ts:

- send: sendTelegramMessage, sendTelegramPhoto, sendWelcomeMessage,
  sendMarveenAvatarChange, sendAvatarChangeMessage, sendMarveenAlert
- read: readAgentTelegramConfig, readMarveenTelegramConfig,
  parseTelegramToken
- validate / cache: validateTelegramToken, marveenBotUsernameCache,
  refreshMarveenBotUsername

web.ts: 4729 -> 4564 lines. No behavior change.
